requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart
python-dotenv

# OpenAI and AI/ML
openai>=1.70.0
faiss-cpu>=1.8.0

# Data processing
numpy>=1.26.0
scipy>=1.15.0
scikit-learn>=1.6.0
pandas

# General utilities
requests>=2.32.0
pydantic>=2.11.0
packaging>=24.0